Updated the user's balance after making a transaction as well as the counterparty's.
Implemented the exception in case the user wants to make a transaction to himself.

Implemented an exception in case the user wants to make a transaction but his amount is less than the amount he wants to transact.

+++ src/main/java/com/bank/bankingapp/entity/Transactions.java
@@ -30,7 +30,7 @@ public class Transactions {
     @JoinColumn(name = "username" , referencedColumnName = "username")
     private User user;
     @Column(name = "suma" , nullable = false)
-    private Long suma;
+    private int suma;
     @Column(name = "moneda" , nullable = false)
     private String moneda;
     @Column(name = "message")
